+++ samplesheets/admin.py
@@ -1,3 +1,12 @@
 from django.contrib import admin
 
-# Register your models here.
+from .models import Investigation, Study, Assay, GenericMaterial, Protocol, \
+    Process
+
+
+admin.site.register(Investigation)
+admin.site.register(Study)
+admin.site.register(Assay)
+admin.site.register(GenericMaterial)
+admin.site.register(Protocol)
+admin.site.register(Process)