+++ samplesheets/urls.py
@@ -8,4 +8,9 @@ urlpatterns = [
         view=views.ProjectSheetsView.as_view(),
         name='project_sheets',
     ),
+    url(
+        regex=r'^(?P<project>\d+)/import$',
+        view=views.SampleSheetImportView.as_view(),
+        name='sheet_import',
+    ),
 ]